better stream writing in HTTPClient FS#2036
This changes the HTTP stream to blocking while writing to the stream
using select() to handle timeouts. Addtionally, wwriting is done in 4k
block now (as it is done with reading).

This is supposed to fix a problem with writing to a SSL stream that is
not quite ready.

Reading from the stream continues to be non-blocking as before.