Use Base64 encoding for long subjects FS#2169
Quoted-Printable specifies a maximum line length and some mail tools
(Apple mail and Thunderbird) take this quite serious and will fail to
decode subjects encoded with quoted-printable when the subject exceeds
the length limit.

The correct fix would be to wrap the header into multiple lines. But
this seems not to be possible with mails() $subject variable.

This patch switches to Base64 encoding for long subjects.

A general decision if switching completely to Base64 is the best way
to go is still open. (see bugreport)