Temporary conversion plugin for new fnencode FS#2197
This plugin converts existing filenames that were stored using the
option "safe" in fnencode from using the dot (.) as post_indicator to
using a bracket (]) as post_insdicator. It will also add a
post_indicator at the end of the file name should it be missing (Bug
FS#2122).

This plugin needs testing by people using the safe encode option!