correctly disable unfinished media options popup
We disabled the media options shortly before the last release because of
an unfixed bug. The disabling was not complete which caused problems
with templates not incorporating the needed CSS.
Since the option dialog hasn't been fixed yet, this patch also disables
the creation of the corresponding HTML to avoid the mentioned template
problems.