Added 3way merge from PEAR Text_Diff3
This adds 3way merge functionality to our DifferenceEngine stack. The
code was taken from the PEAR Text_Diff3 package and adapted to work with
our implementation.

For the curious:
Here's a short Origin tree of the different versions of the Diff code to
make clear how our implementation is related to the PEAR package.

-- original code Geoffrey T. Dairiki
   `-- Adaption for the Horde project
       |-- Adaption for phpwiki
       |   `-- Adaption for MediaWiki
       |       `-- Adaption for DokuWiki
       `-- Creation of PEAR Text_Diff