Add line endings at the end of the file
The bug that is fixed here may have corrupted your search index in a way
that it produces wrong or missing results and won't be fixed
automatically. This occurs when you have deleted the last occurrence of
a word that has been on the last line of one of the word indexes. A
functionality for checking for a broken search index will be added.

The index can be fixed by deleting it completely (remove all .idx files
in data/index/) and recreating it using bin/indexer.php -c. The
searchindex plugin will be updated to be able to do the same, soon.