Deprecation marker for JavaScript functions
This adds a DEPRECATED() JavaScript function. This function will print a
warning to the Browser's debug console if available (Chrome and Firefox
with Firebug extension) when ever it is called.

The DEPRECATED() function was also added to the $() function which
should no longer be used and be replaced with JQuery calls.

Other deprecated functions need to be identified and marked.