Fix the permission and file existence check in the recent changes
Before it was possible that changes for media files the user can't
access were shown if the user could access a page with the same name.
This also reverts the file existence check so it now really checks for
media changes that the media file exists and for page changes the page
and that not one of the two suffice for both.