fix namespace cleaning FS#2295
This patch applies the correct pagename cleaning to namespaces as well.
Namespaces should follow the same rules as pagenames but due to a bug it
was possible to have namespaces ending in a underscore.

If you used such a namespace it will be inaccessible after applying this
patch. You should rename the namespace directory on the filesystem
(removing the trailing underscore). You old links will then work
automatically again. Rebuilding the search index is recommended.