Fixes a problem with parsing overlong changelog lines
When an overlong edit summary was given for an edit, the resulting
changelog line could be longer than the chunk that is handled in the
changelog reader (8192 bytes) causing the reader to abort the operation.

This meant that old revisions where no longer accessible. This patch
fixes the reader (it continues reading chunks until a full line is
found).

However, limiting the summary makes sense and will be added in another
patch.