Added HTML wrapper for mails
The mailwrap.html adds the standard headers and footers to all HTML
mails. The signature present in the text body will be removed before
inserting in the wrapper, allowing a nicer footer for HTML.

Users can overwrite the file with their own to create HTML mails in
corporate design. However, a way to automatically embed referenced
images is missing currently.