allow image embeds in HTML mail templates
Using the place holder @MEDIA(<mediaid>)@ you now can directly embed
images from the media repository into your HTML mail template.

An example was added to the mailwrap.html, but because images are
embedded as is (no prescaling) this is suboptimal. If we keep it, a
smaller version of the DokuWiki logo should be shipped with DokuWiki.