Using sensefull error codes.
Since there are currently a lot of error coded returning 1, that mean
completly different thing, i guess it would be very nice to change this.
A client should not be forced to parse the error message, the error code
should be enough to explain the error.

This change suggests some error codes, that have a hierarchical
structure. In the following list the categories begin with = and the
error codes actually used with -.

= 100 	Page errors
== 110	Page access errors
--- 111	User is not allowed to read the requested page
--- 112	User is not allowed to edit the page
== 120	Page existance errors
--- 121	The requested page does not exist
== 130 	Page edit errors
--- 131	Empty page id
--- 132	Empty page content
--- 133	Page is locked
--- 134	Positive wordblock check

= 200 	Media errors
== 210	Media access errors
--- 211	User is not allowed to read media
---	215	User is not allowed to list media
== 220	Media existance errors
--- 221	The requested media does not exist

= 300	Search errors
== 310	Argument errors
--- 311	The provided value is not a valid timestamp
== 320	Search result errors
--- 321	No chances in specified timeframe