added (failing) test for cleanText()
The cleanText function is supposed to convert DOS to Unix lineendings
but it seems that it doesn't always do that correctly as this thread
suggests: http://forum.dokuwiki.org/thread/8141

I added a unit test that currently fails but haven't found the real
cause yet. Further testing (and a fix) is needed.