Disable some parser and renderer caches during tests FS#2564
This disables some caches which use static variables in order to prevent
parsing or rendering of the same page more than once during a run and
also disables the cache of parser modes. This means we can no longer
test this behavior but at least it fixes syntax plugin tests and tests
which change pages in order to test certain caching behaviors.