Add support for named templates
This allows you to use named templates instead of only [_]_template.txt.
The main benefit is if you have many sub-namespaces each of which has
the same page.

A summary page on movies, for example. A namespace for movies could
contain __summary.txt and then a subnamespace for each movie could be
created. In each movie's namespace, the summary page would be able to
pull from the template the the parent directory.