Load plugins using include_once instead of include
The problem with using include is that when the file that is loaded
does not contain the plugin class (e.g. because the directory name is
wrong), the file could be loaded again when the plugin class is
requested a second time which will lead to a fatal error because of
class redeclaration.