Fix content of pages with incorrect metadata in list feeds
Before this change the last modification time was taken from metadata
and then the revision that was specified there was loaded from the
attic. Now revisions from attic are only loaded when the page has been
modified more recently and the date is normally the last modification
time of the page source. This was a problem when the most recent change
was an external edit, then an older revision was shown in the feed.