re-added linear gradients for firefox
We removed them because of FS#2447 a while ago. But as the original
problem was most probably only caused by the gradient for the page
background and that's using a PNG now, the gradients for the buttons
should be fine. Please test to double check if this change causes
performance problems in Firefox.