disabled JavaScript compatibility layer
this disables the JavaScript compatibility layer which let pre-jQuery
migration plugins run with their old code.

This will break plugins using outdated JavaScript. These plugins have to
be updated according to https://www.dokuwiki.org/devel:jqueryfaq

As a courtesy to users, the compatibility.js is not deleted, yet.
Undoing this single patch can be used as a temporary workaround.