make sure loaded text has the right encoding
When pages contain non-UTF8 chars (eg. when posted through a script or
when edited on the filesystem, parts of DokuWiki can break resulting in
missing page content. This fixes these problems by forcing the content
to UTF-8 on load. This will result in bad characters for input that is
not latin1 but contents will at least be visible.