Indexer: Remove broken and dead readdircache code FS#2771
The code that is removed in this commit has either never been used
(listIndexLenghts) or was completely broken (cacheIndexDir) and was
introduced in the indexer rewrite in 2010. The idea of the rewrite was
to update the readdir cache after every index change instead of on demand.
What the code actually did was removing every updated index from the
cache as it used a wrong if condition. Simply fixing the condition
wouldn't fix the problem as then only updated indexes would be added to
the cache and furthermore the rewrite simply ignored the readdircache
setting. For now the safest solution seems to be removing the code. It
could be added again in a changed form in a future version.