In debug mode: catch and log JS errors with file information
When debugging is enabled (allowdebug enabled) JS errors are now catched
for each non-core JS file and logged with the additional information
from which file they came. This should make it easier to find out which
plugin is the cause for broken JS code. The feature isn't enabled by
default as defining functions inside try-clauses isn't allowed in strict
mode and causes a warning at least in Firefox.