Revert changes to phpseclib to previous commit
Michael Hamann commented in pull request #306 that phpseclib has not
been adopted by the DokuWiki project. Therefore, the changes made to
reduce the CodeSniffer violations should be reverted and the local copy
of phpseclib should be excluded from CS checks. This commit reverts the
previous changes and adds the appropriate exclusion rule.