Removed old default template
The old 'default' template was deprecated with the release of the new
'dokuwiki' template (2012-10-13). It is still available in a separate
repository here: https://github.com/dokufreaks/template-default

But no-one in the core development team intends to maintain this template
anymore. So, anyone willing to adopt it is welcome to clone it and make it
your own.