Improved pagetools event in dokuwiki template
This moves the template name away from the hook name into the event data (as
the 'tpl' value). Therefore it moves the previous data items into the 'items' sub array.
A 'view' value has also been added to describe in which template file the
event was called ('main' or 'detail').

This way it will be much easier for plugins to consider default cases and
differentiate between main and other layouts.