Prepare an XMLRPC logoff method. Rationale: XMLRPC is thought for script, which typically should not store their authentication cookies on disk. However: 1st: DW cannot tell if an external script is in this respect well-behaved. 2nd: it does not hurt to provide means for a voluntary log-out.
There are, BTW, two parts of authentication data: one is stored in the
cookie-storage of the client, and the other part is stored in the
session data of the DW instance on the server. This logoff call is
responsible for invalidating the credentials stored on the server,
regardless of any cookie data remaining (or being stolen) on the client
side.