remove duplicate plugin code for syntax plugins
This makes Doku_Parser_Mode inherit from DokuWiki_Plugin which allows
for the removal of a bunch of duplicate code form
DokuWiki_Syntax_Plugin. This makes the code easier to maintain and makes
sure all DokuWiki plugins are actual instances of DokuWiki_Plugin.

However this adds a bunch of functions to the "normal" parser modes that
don't need them which could have performance/RAM implications.