fix renderer:plugin() signature
in 82d616353e4c3680d88f083eb6f88fe68de92904 we introduced passing the
raw match for syntax plugins to the plugin method of the renderer. This
was introduced for renderer plugins that might need access to the raw
syntax (like the table editor or WYSIWYG plugins). However the function
signature was never updated, making these additional parameter basically
secret. With strict standard this throws errors. This patch fixes this.