Add FULLTEXT_PHRASE_MATCH event for allowing plugins to match phrases
Our index doesn't support phrase searches so we are searching for the
pages that contain all words of the phrase and then search again in the
content of the pages. As plugins can also add additional text to the
index this event allows plugins to do phrase matching in their content.