diff of removed page, require handling right rev=0
When page is removed, and diff can be requested between a revision and
current situation. This results in right revision is 0.

Similar case just after creating a page. A diff between the first
version and nothing before, result in left revision is 0.

In these cases a empty dummy revision is placed as selected value in
dropdown. Otherwise user got distracted by the revisions details shown
in select field,  which are not related to the diff below.