allow config presets with installer
this adds the ability to place a install.conf file next to the
install.php which contains additional config options to be written to
conf/local.php on completion of the installer. The install.conf is
automatically deleted then.

This is useful for automated DokuWiki installers or downloaders that
take care of downloading DokuWiki and then drop the user in the default
install dialog for initial setup. These tools may set up host specific
things options like rewrites, image magic path or file permission
settings in this preset.