allow for easier HTTPClient debugging
Sometimes users report bugs or problems releated to HTTP connections
made by our HTTPCient (eg. in the extension manager or XMLRPC).
Debugging those problems previously required modifying some sources to
enable the debugging output of the HTTPClient.
This patch adds a way to enable the debugging via a request parameter.
Since HTTP requests are often done only *after* some button was clicked,
the parameter is also looked for in the referrer of a request.
Debugging output is enabled through the existence of the parameter
'httpdebug' and is directly printed to the screen. Debugging HTTP only
works when allowdebug is enabled.