The cache should create a new script
The script should be regenerated whenever the JS_CACHE_USE says so, the
content of the script files change or if the amount of the script file
changes. This could be triggered by the new JS_SCRIPT_LIST event. When
the list does not change, the cache will be reused.