send JavaScript with correct mimetype
While Browsers (IE of course) still fail to accept the correct
application/javascript mimetype in the type attribute of the script
element, we should serve the scripts with the correct Content-Type
header at least. This is especially important as the default
configuration of mod_deflate expects application/javascript and will not
compress text/javascript.