Let extension a chance to run a post install script
With this patch, when we install a plugin or a template, if it contains
a file post_install.php, it is executed just after installation.

The use case is: I have tests for my plugin, in the same repo. Hence
installation from github also copies and keeps those files. For 200ko
of useful files, there are 616Ko of file not needed in production.

If I can run a post installation script, I can remove them.

I believe it could also be used to bust caches for plugins coming with
js or css.