a first go at the admin styles
I'm not so sure about using the colors here. On the one hand they help
distinguishing the different topics and our current icons are colored,
too. On the other hand it might be too out of color scheme...

Note this will break with the next update of the icon font until
https://github.com/Templarian/MaterialDesign/issues/644 is fixed. If
this does not get fixed we will have no choice but to include the lookup
CSS as well, though I would prefer if we could avoid it.