some refactoring in the form building
Now the whole form is inserted as a single HTML element. This makes it
easier to correctly position it in the form (the position is determined
dynamically now) than to have to work with permanently shifting
positions witin the form object. Since we don't expect other plugins to
modify our form this shuld be okay.