restyle the inpage output #20 SPR-320
The data is now added as a new section and the table styled more
minimalistic.

This works, however people might not like it. Also this adds a new
section editing button that loads an empty editor. On the other hand it
makes sure the real section before has it's own edit button.

How all this works on export is still untested and probably broken.