use new heuristic to decide on rendering output
To avoid that our output is added to instructions that are not the main
page's instructions we now check what wiki page was read from disk last.
We only render on instruction creation immeadiately after that event and
ignore all others. This should also allow us to ignore the current
action when deciding if we want to add to the instructions.

We have to see if this works out in real life.