implement @retry annotation for our test suite
Some of our tests rely on external resources (like HTTP servers) that
are somewhat flaky sometimes. Instead of having them fail whenever
something goes wrong at the other end, it would be nice to be able to
just retry them a few times before really failing.

This implements a @retry annotation that allows a test to be repeated
the given number of times before it is marked as failed.