integrate select() into SchemaData
Because the signature of buildGetDataSQL() had to change a lot of tests
are breaking currently. Those tests need to be adjusted first, then we
can see if the implemention itself is correct.

After this has been done, the select has to be adjusted to select the
pageid as PID to make it available to possible select() implementations
that rely on the availability of that alias.