allow multicolumn filters #124
when referencing a multi value column through the $STRUCT.<column>$
filter placeholder, all found values are used for the comparison using
an OR sub group.

To make this work some changes had to be introduced. You can only
reference _one_ column in a filter, not multiple. Anything before or
after the column placeholder will be applied to each value in the
multicolumn.

Types overrifding the filter() method need to make sure they handle
arrays in $value correctly.