Initialize DateType with dformat configuration
We're not using strftime in our Date and DateTime types because we want
to be able to represent dates outside the unix timestamp range.

To reuse the strftime format string we have in DokuWiki it needs to be
converted to a date() compatible format. The class added for that could
ultimately be moved to a third party project.