Trigger an event when the state of an extension changes
This triggers an informational event whenever the state of an extension
is changed through the extension manager.

The passed data contains a helper_plugin_extension_extension object and
the type of action that has been executed (install, update, uninstall,
disable, enable).

Plugins could use this to execute things when something is installed, or
plugins are disabled.

The actual use case is that the farmer plugin would like to deactivate
plugins by default when they are installed in the farmer. Plugins could
then be activated on a case-by-case base for each animal.

Theoretically it could also be used by plugins to install their own
dependencies, but I guess that would be better implemented as a
functionality of the extension manager itself.