apply fixed-width font to uses of the <tt> element
Monospaced text is usually used in the context of a code fragment or embedded text file. In both of these cases it's desirable to have the content set inside a slightly off-colored background. At other times it would be useful to format a text fragment to use a fixed-width font without affecting anything else about it. The 'creole' plugin uses the <tt> tag for this purpose by having users wrap their text within "##". Without any CSS styling, the result does not look very good.

This commit pulls the 'font-family' CSS selector out of 'pre', 'code', 'samp', and 'kbd' rules and into a mixin, which is then applied to a new 'tt' rule.