refactor SchemaData into AccessTable with factory
first part of the refactoring for the SchemaData and SchemaLookupData
classes. Both now inherit from AccessTable which provides common methods
and two factory methods to automatically create the correct class
depending on the schema.

Both classes no longer inherit from Schema. Instead they contain a
Schema as a member variable. This makes it possible to initialize them
with an existing schema (which is required for the factory methods).

The flow is now like this:

1: init a schema (read info from the database)
2: look at the schema to figure out if it's a lookup or not
3: initialize the right Data object with the schema