Check the server has a sensible time
DokuWiki assumes that the server's time is correct. Especially page
revisions and cache handling depend on correct time. If that's not the
case it can lead to problems later (as mentioned in #1644).

This patch adds a very simple time check using the Date response header
from the DokuWiki server to our do=check mechanism.