Add Lookup Support to Dropdown
Instead of specifying values directly, one can now specify a lookup
schema where the data is read from. The specified field name may contain
a $LANG place holder to be replaced with the current language string.

There is currently no enforcement that the specified schema is indeed a
lookup schema. It should work with a non-lookup as well.