AuthPDO: allow in-SQL pasword checking
This is a feature that was present in the old authmysql plugin. It
allows administrators to specify an SQL string that is used for explicit
password checking instead of letting DokuWiki compare the hashes. This
is useful where application implement their own salt/hashing thing.

This comes with a test case for the MyBB forum software. Dump kindly
provided by William Kitchen