first start at refactoring the validation mechanism
A Validator is now returned by the AccessTable. It has it's own save
method. The entry.php was split up in entry, edit and revert to separate
the different concerns. Validation is only done in entry and the thanks
to the new Validator mechanism, the once loaded access table is reused
in saving.

This will break all kind of stuff. Fixing that is for the upcoming
commits