This adds a new type Color
This is not terribly useful on it's own, but will be used by other
plugins defining their own types and making use of this.

The default color will never be saved.

This type makes use of <input type="color"> which is not supported in
some older browsers (a simple text input will be displayed by those).