feat: store user and rev of all pages
Rename `meta/Title.php` `meta/Page.php` since it now handles the user
and timestamp for the last change as well.

This is done in preparation to adding a %lasteditor% selector showing
the last editor of a page and changing the %lastupdated% to show the
time when the page was last changed instead of showing the time when the
struct data of the page was last updated.

SPR-622